Correcciones en los servicios Añadido ArticuloService

+++ src/main/java/com/ventas/models/BaseModel.java
@@ -13,7 +13,7 @@ public abstract class BaseModel implements Serializable {
         this(0);
     }
 
-    public abstract long getID();
+    public abstract int getID();
 
     public boolean isDelete() {
         return isDelete;
